# server/docs/swagger.yaml
basePath: /
definitions:
  request.FileInfo:
    properties:
      content:
        type: string
      login:
        type: string
      sha:
        type: string
      token:
        type: string
      uid:
        type: string
    type: object
  response.FileInfo:
    properties:
      content:
        type: string
      download_url:
        type: string
      html_url:
        type: string
      name:
        type: string
      path:
        type: string
      sha:
        type: string
      size:
        type: integer
      type:
        type: string
      uid:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server Petstore server.
  title: ShowNote API
  version: 0.0.1
paths:
  /repo/file:
    delete:
      description: 删除文件
      parameters:
      - description: access_token
        in: query
        name: token
        required: true
        type: string
      - description: 用户名
        in: query
        name: login
        required: true
        type: string
      - description: 文件UID
        in: query
        name: uid
        required: true
        type: string
      - description: 文件sha
        in: query
        name: sha
        required: true
        type: string
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 删除文件
      tags:
      - 仓库
    get:
      description: 获取仓库具体路径下的内容
      parameters:
      - description: access_token
        in: query
        name: token
        required: true
        type: string
      - description: 用户名
        in: query
        name: login
        required: true
        type: string
      - description: 文件UID
        in: query
        name: uid
        required: true
        type: string
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 获取仓库具体路径下的内容
      tags:
      - 仓库
    post:
      description: 新建文件
      parameters:
      - description: FileInfo
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/request.FileInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 新建文件
      tags:
      - 仓库
    put:
      description: 更新文件
      parameters:
      - description: FileInfo
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/request.FileInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 更新文件
      tags:
      - 仓库
  /repo/info:
    get:
      description: 获取仓库信息，没有则初始化仓库，并返回.shownote/workspace.json的内容
      parameters:
      - description: access_token
        in: query
        name: token
        required: true
        type: string
      - description: 用户名
        in: query
        name: login
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 获取仓库信息
      tags:
      - 仓库
  /repo/upload:
    post:
      description: 上传文件
      parameters:
      - description: 文件类型,image|viodo|audio
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 上传文件
      tags:
      - 仓库
  /repo/workspace:
    get:
      description: 获取workspace内容
      parameters:
      - description: access_token
        in: query
        name: token
        required: true
        type: string
      - description: 用户名
        in: query
        name: login
        required: true
        type: string
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 获取workspace内容
      tags:
      - 仓库
    put:
      description: 更新workspace
      parameters:
      - description: FileInfo
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/request.FileInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 成功后返回值
          schema:
            $ref: '#/definitions/response.FileInfo'
      summary: 更新workspace
      tags:
      - 仓库
  /user/auth/:repo:
    get:
      description: gitee auth login
      parameters:
      - description: 仓库类型 gitee or github
        in: query
        name: repo
        required: true
        type: string
      - description: '{redirect_uri}?code=abc'
        in: query
        name: code
        required: true
        type: string
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      summary: 授权登录
      tags:
      - 用户
  /user/info:
    get:
      description: 获取用户的基本信息
      parameters:
      - description: accesstoken
        in: query
        name: token
        required: true
        type: string
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      summary: 获取用户信息
      tags:
      - 用户
swagger: "2.0"